feat(library): simplify tagging panel + structured live tagger progress
The /admin/library tagging panel had accreted three stacked progress
meters with long explainers, while live-run feedback was a client-side
coverage-delta guess that broke on page reload.

Backend:
- new music/tagger-progress.ts: [progress] stdout sentinel + shared
  TaggerProgress type; reportProgress() emitted from every pipeline
  phase of tag-library.ts (walk/enrich/embed/seed/propagate/learn/done)
  and the shared analysis pass (covers the standalone analyzer too)
- broadcast/tagger.ts parses sentinels into tagger.progress with
  per-stream line buffering (also fixes chunk-split log lines) and
  keeps them out of lastLog; rides on /settings for free
- LLM phases report per-leg dual-LLM counts and failed-track counts

Web:
- mood & energy is the single hero meter; acoustic bpm/key and audio
  fingerprint rows (meters, toggle, Analyze) move into the renamed
  'Maintenance & advanced' drawer with trimmed copy
- running view shows the real phase label, x/y counts, percent bar
  (indeterminate pulse during the walk), active-learn round, dual-LLM
  per-leg counts and error count — and survives page reloads since
  progress now comes from the server (runInfo baseline hack removed)
- TaggingPanel extracted to LibraryTaggingPanel.tsx (mechanical split)

Also untracks controller/node_modules — an accidentally-committed
broken symlink to a macOS-local path (slipped past .gitignore's
node_modules/ dir pattern in #325).
