--- controller/node_modules
@@ -1,1 +1,0 @@
-/Users/klair/Projects/subwave/web/controller/node_modules
